Move TokenExchange back into the weco.http.client namespace
So we can use it in the METS adapter/transformer for the storage service
client.
