+++ RELEASE.md
@@ -1,0 +1,3 @@
+RELEASE_TYPE: minor
+
+Move TokenExchange back into the weco.http.client namespace.
